feat: update deps and move types to @types
