+++ package.json
@@ -30,31 +30,27 @@
   "author": "Anthony Yarbrough",
   "license": "MIT",
   "devDependencies": {
-    "@babel/core": "^7.6.2",
-    "@babel/plugin-syntax-jsx": "^7.2.0",
-    "@babel/types": "^7.6.1",
-    "@commitlint/cli": "^8.1.0",
-    "@commitlint/config-conventional": "^8.1.0",
+    "@babel/core": "^7.8.4",
+    "@babel/plugin-syntax-jsx": "^7.8.3",
+    "@babel/types": "^7.8.3",
+    "@commitlint/cli": "^8.3.5",
+    "@commitlint/config-conventional": "^8.3.4",
     "@types/babel-plugin-syntax-jsx": "^6.18.0",
     "@types/babel__core": "^7.1.3",
-    "@types/jest": "^24.0.18",
+    "@types/jest": "^25.1.1",
     "@types/lodash.camelcase": "^4.3.6",
-    "@typescript-eslint/eslint-plugin": "^2.3.1",
-    "@typescript-eslint/parser": "^2.3.1",
-    "babel-eslint": "^10.0.3",
-    "eslint": "^6.4.0",
-    "husky": "^3.0.5",
-    "jest": "^24.9.0",
-    "prettier": "^1.18.2",
-    "rollup": "^1.21.4",
+    "husky": "^4.2.1",
+    "jest": "^25.1.0",
+    "prettier": "^1.19.1",
+    "rollup": "^1.31.0",
     "rollup-plugin-progress": "^1.1.1",
-    "rollup-plugin-sourcemaps": "^0.4.2",
-    "rollup-plugin-terser": "^5.1.2",
-    "rollup-plugin-typescript2": "^0.24.2",
-    "standard-version": "^7.0.0",
-    "ts-jest": "^24.1.0",
+    "rollup-plugin-sourcemaps": "^0.5.0",
+    "rollup-plugin-terser": "^5.2.0",
+    "rollup-plugin-typescript2": "^0.25.3",
+    "standard-version": "^7.1.0",
+    "ts-jest": "^25.1.0",
     "tslib": "^1.10.0",
-    "typescript": "^3.6.2"
+    "typescript": "^3.7.5"
   },
   "husky": {
     "hooks": {
@@ -67,7 +63,7 @@
     ]
   },
   "dependencies": {
-    "known-css-properties": "^0.16.0",
+    "known-css-properties": "^0.18.0",
     "lodash.camelcase": "^4.3.0"
   }
 }
