+++ src/mergers.ts
@@ -1,5 +1,6 @@
 import { types as t } from '@babel/core'
 import { ArrayExpression, ObjectExpression, ObjectProperty } from '@babel/types'
+
 import { buildObjectProperty, buildStyleObject } from './builders'
 
 const mergeCssProps = (
