+++ tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["src", "types"],
+  "include": ["src"],
   "compilerOptions": {
     "target": "es5",
     "module": "esnext",
